+++ tests/test_ogcapi_features_ldproxy.py
@@ -2,7 +2,7 @@ from tests.utils import service_ok
 
 import pytest
 
-from owslib.ogc_api.features import Features
+from owslib.ogcapi.features import Features
 
 SERVICE_URL = 'https://www.ldproxy.nrw.de/rest/services/kataster/?f=json'
 
@@ -11,7 +11,7 @@ SERVICE_URL = 'https://www.ldproxy.nrw.de/rest/services/kataster/?f=json'
 @pytest.mark.skip(reason='api() call fails. See issue #625')
 @pytest.mark.skipif(not service_ok(SERVICE_URL),
                     reason='service is unreachable')
-def test_wfs3_ldproxy():
+def test_ogcapi_features_ldproxy():
     w = Features(SERVICE_URL)
 
     assert w.url == 'https://www.ldproxy.nrw.de/rest/services/kataster/'
